first-build-on

+++ Assets/Resources/Scripts/Utils.cs
@@ -1,6 +1,4 @@
 ﻿using System;
-using System.Collections;
-using System.Collections.Generic;
 using System.IO;
 using System.Runtime.Serialization.Formatters.Binary;
 using UnityEngine;
@@ -9,7 +7,7 @@ namespace Resources.Scripts
 {
     public static class Utils
     {
-        public static readonly float ScreenDif = Display.main.systemWidth / 1280;
+        public static readonly float ScreenDif = (float) Display.main.systemWidth / 1280;
     }
     
     public static class ExtensionMethods
